+++ cogs/modmail.py
@@ -617,9 +617,7 @@ class Modmail(commands.Cog):
                 color=self.bot.error_color, description="Message not found or no longer exists."
             )
         else:
-            embed = discord.Embed(
-                color=self.bot.main_color, description=message.jump_url
-            )
+            embed = discord.Embed(color=self.bot.main_color, description=message.jump_url)
         await ctx.send(embed=embed)
 
     @commands.command()